consolidate hashing logic in utilities
